Cumulative authorization policies
Replaced "SessionBusiness" and "SessionClient" policies with "Business"
and "Client" which have no authentication scheme and only check for bid
and cid claims, respectively. These policies are to be used with the
authentication policies "UserPass" and "Session".

The "UserPass" policy now retrieves the user's bid or cid like the
"Session" policy so it supports the "Business" and "Client" policies.