Web Client Edit Account Info
Form is pre-populated by calling the APIs getinfo method in Page_Load. Client must enter current username and password (UserPass) for changes to take effect. We should consider renaming this "ClientEditAccountInfo", since the fields and are those of client only accounts. In the future, another page will need to be made for businesses (OR can improve api to determine which fields are necessary by user_type).